# resources/swagger/swagger.yml
info:
  title: Comentario
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
swagger: "2.0"
basePath: /api
securityDefinitions:
  # Header authentication for commenters (used by the embed part)
  userSessionHeader:
    type: apiKey
    in: header
    name: X-User-Session

  # Cookie authentication for users (used by the frontend part). Uses the apiKey type for the lack of a proper
  # cookie-based authentication in Swagger 2
  userCookie:
    type: apiKey
    in: header
    name: Cookie

  # Token authentication for certain endpoints, called externally, such as password reset or email confirmation
  token:
    type: oauth2
    flow: accessCode
    authorizationUrl: http://dummy/
    tokenUrl: http://dummy/
    scopes:
      confirm-email: confirm user's email
      login: authenticate the user
      pwd-reset: reset user's password

# Default security is cookie-based user authentication
security:
  - userCookie: []
definitions:
...
  dynamicConfigItem:
    description: Dynamic configuration item
    type: object
    required:
      - key
      - value
    properties:
      key:
        type: string
        maxLength: 255
        description: Item key
      value:
        type: string
        maxLength: 255
        description: Item value
      datatype:
        $ref: "#/definitions/dynamicConfigItemDatatype"
        readOnly: true
        description: Item datatype
      updatedTime:
        type: string
        format: date-time
        readOnly: true
        description: Timestamp when the item was last updated in the database
      userUpdated:
        type: string
        format: uuid
        readOnly: true
        description: Reference to the user who last updated the item in the database
      defaultValue:
        type: string
        readOnly: true
        description: Item's default value
      section:
        type: string
        description: Key of the section the item belongs to
      min:
        type: integer
        readOnly: true
        description: Minimum allowed value of the setting
        x-omitempty: false
      max:
        type: integer
        readOnly: true
        description: Maximum allowed value of the setting
        x-omitempty: false
  dynamicConfigItemDatatype:
    description: Dynamic configuration item datatype
    type: string
    enum:
      - bool
      - int
...
  federatedIdpId:
    description: Federated identity provider ID
    type: string
    maxLength: 37
    pattern: '^facebook|github|gitlab|google|twitter|(oidc:[-a-z0-9]{1,32})$'
    x-isnullable: false
...
  releaseMetadata:
    description: Release information
    type: object
    required:
      - name
      - version
      - pageUrl
    properties:
      name:
        type: string
        description: Release name
        x-isnullable: false
      version:
        type: string
        description: Released version
        x-isnullable: false
      pageUrl:
        type: string
        description: Release page URL
        x-isnullable: false
...
parameters:
...
  pathUuid:
    in: path
    name: uuid
    required: true
    description: UUID in the path
    type: string
    format: uuid
    x-isnullable: false
...
  queryToken:
    in: query
    name: token
    required: true
    description: ID consisting of 64 hex digits, in the query parameter
    type: string
    minLength: 64
    maxLength: 64
    pattern: '[0-9a-f]{64}'
...
paths:
...
  /e2e/config/dynamic:
    put:
      operationId: E2eConfigDynamicUpdate
      summary: Update the given dynamic instance configuration items
      tags:
        - ApiE2e
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/dynamicConfigItem"
      responses:
        204:
          description: Dynamic instance configuration item has been updated
  /e2e/config/versions/latestRelease:
    put:
      operationId: E2eConfigVersionLatestReleaseUpdate
      summary: Set the latest release returned by the app
      tags:
        - ApiE2e
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/releaseMetadata"
      responses:
        204:
          description: Dynamic instance configuration item has been updated
  /e2e/domains/{uuid}:
    parameters:
      - $ref: "#/parameters/pathUuid"

    patch:
      operationId: E2eDomainPatch
      summary: Patch properties of specified domain. The body must be JSON with domain properties to patch
      tags:
        - ApiE2e
      security: []
      responses:
        204:
          description: Domain properties have been updated
  /e2e/domains/{uuid}/config:
    put:
      operationId: E2eDomainConfigUpdate
      summary: Update the given dynamic configuration item values for the specified domain
      tags:
        - ApiE2e
      security: []
      parameters:
        - $ref: "#/parameters/pathUuid"
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/dynamicConfigItem"
      responses:
        204:
          description: Domain configuration item has been updated
  /e2e/domains/{uuid}/idps:
    put:
      operationId: E2eDomainUpdateIdps
      summary: Update the list of federated IdPs of specified domain
      tags:
        - ApiE2e
      security: []
      parameters:
        - $ref: "#/parameters/pathUuid"
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/federatedIdpId"
            description: IDs of federated identity providers enabled for the domain
      responses:
        204:
          description: Domain's IdP list has been updated
...
  /e2e/oauth/{uuid}/sso/noninteractive:
    get:
      operationId: E2eOAuthSSONonInteractive
      summary: Perform a fake non-interactive SSO-auth for a domain with the given UUID
      tags:
        - ApiE2e
      security: []
      parameters:
        - $ref: "#/parameters/pathUuid"
        - $ref: "#/parameters/queryToken"
        - in: query
          name: hmac
          required: true
          type: string
          minLength: 64
          maxLength: 64
          pattern: '[0-9a-f]{64}'
          description: HMAC signature for token
      responses:
        302:
          description: Auth successful, redirecting to the callback URL
          headers:
            Location:
              type: string